+++ CHANGELOG.md
@@ -8,6 +8,11 @@
 - Remove Ash ModInitializers completely
   1. Mods no longer have to register themselves with Ash API on NeoForge
   2. NeoForge now allows for client-only entry points
+- Rename ClientTickEvents to ClientTickEvent
+  - Rename start and end callbacks to pre and post to match NeoForge
+- Rename `ModOptionsScreenRegistry#registerModOptionsScreen` to just `ModOptionsScreenRegistry#register`
+- Rename `KeyMappingRegistry#registerKeyMapping` to just `KeyMappingRegistry#register`
+- Rewrite UseBlockEvent to use root level functional interface
 
 ## v20.6.2
 ### v20.6.2-beta